// src/lib/scoretrakapis/scoretrak/proto/v1/uuid_pb.d.ts
import * as jspb from "google-protobuf";

export class UUID extends jspb.Message {
  getValue(): string;
  setValue(value: string): UUID;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): UUID.AsObject;
  static toObject(includeInstance: boolean, msg: UUID): UUID.AsObject;
  static serializeBinaryToWriter(
    message: UUID,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): UUID;
  static deserializeBinaryFromReader(
    message: UUID,
    reader: jspb.BinaryReader
  ): UUID;
}

export namespace UUID {
  export type AsObject = {
    value: string;
  };
}