// src/lib/scoretrakapis/scoretrak/property/v1/property_pb.d.ts
export class Property extends jspb.Message {
  getServiceId(): scoretrak_proto_v1_uuid_pb.UUID | undefined;
  setServiceId(value?: scoretrak_proto_v1_uuid_pb.UUID): Property;
  hasServiceId(): boolean;
  clearServiceId(): Property;

  getKey(): string;
  setKey(value: string): Property;

  getValue(): google_protobuf_wrappers_pb.StringValue | undefined;
  setValue(value?: google_protobuf_wrappers_pb.StringValue): Property;
  hasValue(): boolean;
  clearValue(): Property;

  getStatus(): Status;
  setStatus(value: Status): Property;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Property.AsObject;
  static toObject(includeInstance: boolean, msg: Property): Property.AsObject;
  static serializeBinaryToWriter(
    message: Property,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): Property;
  static deserializeBinaryFromReader(
    message: Property,
    reader: jspb.BinaryReader
  ): Property;
}

export namespace Property {
  export type AsObject = {
    serviceId?: scoretrak_proto_v1_uuid_pb.UUID.AsObject;
    key: string;
    value?: google_protobuf_wrappers_pb.StringValue.AsObject;
    status: Status;
  };
}

export class GetAllRequest extends jspb.Message {
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): GetAllRequest.AsObject;
  static toObject(
    includeInstance: boolean,
    msg: GetAllRequest
  ): GetAllRequest.AsObject;
  static serializeBinaryToWriter(
    message: GetAllRequest,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): GetAllRequest;
  static deserializeBinaryFromReader(
    message: GetAllRequest,
    reader: jspb.BinaryReader
  ): GetAllRequest;
}

export namespace GetAllRequest {
  export type AsObject = {};
}
// ...
export class DeleteRequest extends jspb.Message {
  getServiceId(): scoretrak_proto_v1_uuid_pb.UUID | undefined;
  setServiceId(value?: scoretrak_proto_v1_uuid_pb.UUID): DeleteRequest;
  hasServiceId(): boolean;
  clearServiceId(): DeleteRequest;

  getKey(): string;
  setKey(value: string): DeleteRequest;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): DeleteRequest.AsObject;
  static toObject(
    includeInstance: boolean,
    msg: DeleteRequest
  ): DeleteRequest.AsObject;
  static serializeBinaryToWriter(
    message: DeleteRequest,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): DeleteRequest;
  static deserializeBinaryFromReader(
    message: DeleteRequest,
    reader: jspb.BinaryReader
  ): DeleteRequest;
}

export namespace DeleteRequest {
  export type AsObject = {
    serviceId?: scoretrak_proto_v1_uuid_pb.UUID.AsObject;
    key: string;
  };
}
// ...
export class StoreRequest extends jspb.Message {
  getPropertiesList(): Array<Property>;
  setPropertiesList(value: Array<Property>): StoreRequest;
  clearPropertiesList(): StoreRequest;
  addProperties(value?: Property, index?: number): Property;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): StoreRequest.AsObject;
  static toObject(
    includeInstance: boolean,
    msg: StoreRequest
  ): StoreRequest.AsObject;
  static serializeBinaryToWriter(
    message: StoreRequest,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): StoreRequest;
  static deserializeBinaryFromReader(
    message: StoreRequest,
    reader: jspb.BinaryReader
  ): StoreRequest;
}

export namespace StoreRequest {
  export type AsObject = {
    propertiesList: Array<Property.AsObject>;
  };
}
// ...
export class UpdateRequest extends jspb.Message {
  getProperty(): Property | undefined;
  setProperty(value?: Property): UpdateRequest;
  hasProperty(): boolean;
  clearProperty(): UpdateRequest;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): UpdateRequest.AsObject;
  static toObject(
    includeInstance: boolean,
    msg: UpdateRequest
  ): UpdateRequest.AsObject;
  static serializeBinaryToWriter(
    message: UpdateRequest,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): UpdateRequest;
  static deserializeBinaryFromReader(
    message: UpdateRequest,
    reader: jspb.BinaryReader
  ): UpdateRequest;
}

export namespace UpdateRequest {
  export type AsObject = {
    property?: Property.AsObject;
  };
}
// ...
export enum Status {
  STATUS_UNSPECIFIED = 0,
  STATUS_VIEW = 1,
  STATUS_EDIT = 2,
  STATUS_HIDE = 3,
}